Move uuid5 from bin/loadmeta.jl into src/Types.jl
and use it for the julia UUID and in tests.
